Implemented Github Login WIP
