+++ backend/requirements.txt
@@ -3,6 +3,7 @@ flask-restful
 flask-sqlalchemy
 flask-migrate
 flask-dance
+flask-login
 validator_collection
 python-dotenv
 psycopg2-binary
